Factor pylint path into utils submodule & minor cleanup
